Extract domainWarning from manage page
This ensures the domaim warning can be used from the migration tests
without depending on the management page. Jest cannot load the
management page because it uses next-gen Vite features.
